requirements.txt
asgiref==3.7.2
Django==5.0.3
django-phonenumber-field==7.3.0
psycopg==3.1.18
sqlparse==0.4.4
typing_extensions==4.10.0
phonenumbers==8.13.32
six==1.16.0
environs[django]==10.3.0
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
pillow==10.3.0
boto3==1.34.103
django-storages==1.14.3
django-cleanup==8.1.0
django-imagekit==5.0.0